pyproject.toml
[tool.poetry]
name = "Connect6_Analyze"
version = "0.1.0"
description = ""
authors = ["Tzu-Yee <fdja1974.cs07g@nctu.edu.tw>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
tqdm = "^4.47.0"
numpy = "^1.19.0"
plotly = "4.8.2"
pandas = "^1.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
